Add the "missing" WireGuard rc.d script.
This script reimplements most features of wg-quick(8) in FreeBSD
/bin/sh instead of bash.

The following features have not been reimplemented:
  * The daemon syncing a routing table the AllowedIPs settings.
  * Saving the active configuration back to a config file.

The PreUp/PostUp/PreDown/PostDown hooks are implemented, but the hooks
are executed in sh instead of bash. Addionally the successful execution
is tracked to avoid (re-)execution of hooks because it's expected that users will rerun rc.d scripts.